+++ gherkin/java/src/main/java/io/cucumber/gherkin/GherkinLine.java
@@ -5,8 +5,12 @@ import java.util.List;
 import java.util.PrimitiveIterator;
 import java.util.Scanner;
 
+import static io.cucumber.gherkin.GherkinLanguageConstants.COMMENT_PREFIX;
+import static io.cucumber.gherkin.GherkinLanguageConstants.TAG_PREFIX;
 import static io.cucumber.gherkin.StringUtils.ltrim;
+import static io.cucumber.gherkin.StringUtils.ltrimKeepNewLines;
 import static io.cucumber.gherkin.StringUtils.rtrim;
+import static io.cucumber.gherkin.StringUtils.rtrimKeepNewLines;
 import static io.cucumber.gherkin.StringUtils.symbolCount;
 import static io.cucumber.gherkin.StringUtils.trim;
 
@@ -16,10 +20,12 @@ public class GherkinLine implements IGherkinLine {
     private final String lineText;
     private final String trimmedLineText;
     private final int indent;
+    private final int line;
 
-    public GherkinLine(String lineText) {
+    public GherkinLine(String lineText, int line) {
         this.lineText = lineText;
         this.trimmedLineText = trim(lineText);
+        this.line = line;
         indent = symbolCount(lineText) - symbolCount(ltrim(lineText));
     }
 
@@ -57,7 +63,26 @@ public class GherkinLine implements IGherkinLine {
 
     @Override
     public List<GherkinLineSpan> getTags() {
-        return getSpans("\\s+");
+
+        String uncommentedLine = trimmedLineText.split("\\s" + COMMENT_PREFIX, 2)[0];
+        List<GherkinLineSpan> tags = new ArrayList<>();
+        int indexInUncommentedLine = 0;
+
+        String[] elements = uncommentedLine.split(TAG_PREFIX);
+        for (String element : elements) {
+            String token = rtrim(element);
+            if (token.isEmpty()) {
+                continue;
+            }
+            int symbolLength = uncommentedLine.codePointCount(0, indexInUncommentedLine);
+            int column = indent() + symbolLength + 1;
+            if (!token.matches("^\\S+$")) {
+                throw new ParserException("A tag may not contain whitespace", new Location(line, column));
+            }
+            tags.add(new GherkinLineSpan(column, TAG_PREFIX + token));
+            indexInUncommentedLine += element.length() + 1;
+        }
+        return tags;
     }
 
     @Override
@@ -69,8 +94,8 @@ public class GherkinLine implements IGherkinLine {
         int cellStart = 0;
         boolean escape = false;
         PrimitiveIterator.OfInt iterator = lineText.codePoints().iterator();
-        while (iterator.hasNext() ) {
-            int c = iterator.next() ;
+        while (iterator.hasNext()) {
+            int c = iterator.next();
             if (escape) {
                 switch (c) {
                     case 'n':
@@ -98,9 +123,9 @@ public class GherkinLine implements IGherkinLine {
                         beforeFirst = false;
                     } else {
                         String cell = cellBuilder.toString();
-                        String leftTrimmedCell = ltrim(cell);
+                        String leftTrimmedCell = ltrimKeepNewLines(cell);
                         int cellIndent = symbolCount(cell) - symbolCount(leftTrimmedCell);
-                        lineSpans.add(new GherkinLineSpan(cellStart + cellIndent + OFFSET, rtrim(leftTrimmedCell)));
+                        lineSpans.add(new GherkinLineSpan(cellStart + cellIndent + OFFSET, rtrimKeepNewLines(leftTrimmedCell)));
                     }
                     cellBuilder = new StringBuilder();
                     cellStart = col + 1;
@@ -122,20 +147,4 @@ public class GherkinLine implements IGherkinLine {
                         .equals(GherkinLanguageConstants.TITLE_KEYWORD_SEPARATOR);
     }
 
-    private List<GherkinLineSpan> getSpans(String delimiter) {
-        List<GherkinLineSpan> lineSpans = new ArrayList<GherkinLineSpan>();
-        Scanner scanner = new Scanner(trimmedLineText).useDelimiter(delimiter);
-        while (scanner.hasNext()) {
-            String cell = scanner.next();
-            String leftTrimmedCell = ltrim(cell);
-            int cellIndent = symbolCount(cell) - symbolCount(leftTrimmedCell);
-
-            String trimmedCell = rtrim(leftTrimmedCell);
-            int scannerStart = scanner.match().start();
-            int symbolLength = trimmedLineText.codePointCount(0, scannerStart);
-            int column = 1 + indent() + symbolLength + cellIndent;
-            lineSpans.add(new GherkinLineSpan(column, trimmedCell));
-        }
-        return lineSpans;
-    }
 }