+++ react/javascript/src/components/gherkin/Step.tsx
@@ -3,12 +3,12 @@ import DataTable from './DataTable'
 import Keyword from './Keyword'
 import DocString from './DocString'
 import { messages } from '@cucumber/messages'
-import statusColor from './statusColor'
-import TestResultQueryContext from '../../TestResultsQueryContext'
+import CucumberQueryContext from '../../CucumberQueryContext'
 import UriContext from '../../UriContext'
 import GherkinQueryContext from '../../GherkinQueryContext'
-import StepMatchArgumentsQueryContext from '../../StepMatchArgumentsQueryContext'
 import ErrorMessage from './ErrorMessage'
+import StepContainer from './StepContainer'
+import Attachment from './Attachment'
 
 interface IProps {
   step: messages.GherkinDocument.Feature.IStep
@@ -22,15 +22,13 @@ const Step: React.FunctionComponent<IProps> = ({
   renderMessage,
 }) => {
   const gherkinQuery = React.useContext(GherkinQueryContext)
-  const testResultQuery = React.useContext(TestResultQueryContext)
-  const stepMatchArgumentsQuery = React.useContext(
-    StepMatchArgumentsQueryContext
-  )
+  const cucumberQuery = React.useContext(CucumberQueryContext)
   const uri = React.useContext(UriContext)
 
   const pickleStepIds = gherkinQuery.getPickleStepIds(uri, step.location.line)
-  const pickleStepResults = testResultQuery.getPickleStepResults(pickleStepIds)
-  const testResult = testResultQuery.getWorstResult(pickleStepResults)
+  const pickleStepResults = cucumberQuery.getPickleStepResults(pickleStepIds)
+  const testResult = cucumberQuery.getWorstResult(pickleStepResults)
+  const attachments = cucumberQuery.getPickleStepAttachments(pickleStepIds)
 
   const stepTextElements: JSX.Element[] = []
 
@@ -39,10 +37,10 @@ const Step: React.FunctionComponent<IProps> = ({
       pickleStepIds.length === 0
         ? // This can happen in rare cases for background steps in a document that only has step-less scenarios,
           // because background steps are not added to the pickle when the scenario has no steps. In this case
-          // the bacground step will be rendered as undefined (even if there are matching step definitions). This
+          // the background step will be rendered as undefined (even if there are matching step definitions). This
           // is not ideal, but it is rare enough that we don't care about it for now.
           []
-        : stepMatchArgumentsQuery.getStepMatchArgumentsLists(pickleStepIds[0])
+        : cucumberQuery.getStepMatchArgumentsLists(pickleStepIds[0])
     if (stepMatchArgumentsLists.length === 1) {
       // Step is defined
       const stepMatchArguments = stepMatchArgumentsLists[0].stepMatchArguments
@@ -60,17 +58,14 @@ const Step: React.FunctionComponent<IProps> = ({
         const arg = argument.group.value
         if (arg.length > 0) {
           stepTextElements.push(
-            <span
+            <a
               className="step-param"
               key={`bold-${index}`}
-              style={{
-                backgroundColor: statusColor(testResult.status)
-                  .darken(0.1)
-                  .hex(),
-              }}
+              href="#"
+              title={argument.parameterTypeName}
             >
               {arg}
-            </span>
+            </a>
           )
         }
         offset += plain.length + arg.length
@@ -108,21 +103,21 @@ const Step: React.FunctionComponent<IProps> = ({
   }
 
   return (
-    <li
-      className="step"
-      style={{ backgroundColor: statusColor(testResult.status).hex() }}
-    >
-      <h3>
-        <Keyword>{step.keyword}</Keyword>
-        {stepTextElements}
-      </h3>
-      <div className="indent">
+    <li className="step">
+      <StepContainer status={testResult.status}>
+        <h3>
+          <Keyword>{step.keyword}</Keyword>
+          {stepTextElements}
+        </h3>
         {step.dataTable && <DataTable dataTable={step.dataTable} />}
         {step.docString && <DocString docString={step.docString} />}
-      </div>
-      {renderMessage && testResult.message && (
-        <ErrorMessage status={testResult.status} message={testResult.message} />
-      )}
+        {renderMessage && testResult.message && (
+          <ErrorMessage message={testResult.message} />
+        )}
+        {attachments.map((attachment, i) => (
+          <Attachment key={i} attachment={attachment} />
+        ))}
+      </StepContainer>
     </li>
   )
 }