+++ react/CHANGELOG.md
@@ -9,11 +9,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+* Display attachments (`image/*` and `text/*` media types)
+
 ### Changed
 
 ### Deprecated
 
 ### Removed
+
+### Fixed
+
+## [3.3.0] - 2020-01-22
 
 ### Fixed
 
@@ -78,7 +84,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * First release
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/react/v3.2.0...master
+[Unreleased]: https://github.com/cucumber/cucumber/compare/react/v3.3.0...master
+[3.3.0]:      https://github.com/cucumber/cucumber/compare/cucumber-react/v3.2.0...react/v3.3.0
 [3.2.0]:      https://github.com/cucumber/cucumber/compare/cucumber-react/v3.1.0...react/v3.2.0
 [3.1.0]:      https://github.com/cucumber/cucumber/compare/cucumber-react/v3.0.0...cucumber-react/v3.1.0
 [3.0.0]:      https://github.com/cucumber/cucumber/compare/cucumber-react/v2.0.1...cucumber-react/v3.0.0